Add validations for models & routes

+++ db/schema.rb
@@ -33,7 +33,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_06_20_202225) do
   end
 
   create_table "inventory_foods", force: :cascade do |t|
-    t.integer "quantity", null: false
+    t.float "quantity", null: false
     t.bigint "food_id", null: false
     t.bigint "inventory_id", null: false
     t.datetime "created_at", null: false
@@ -43,7 +43,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_06_20_202225) do
   end
 
   create_table "recipe_foods", force: :cascade do |t|
-    t.integer "quantity", null: false
+    t.float "quantity", null: false
     t.bigint "food_id", null: false
     t.bigint "recipe_id", null: false
     t.datetime "created_at", null: false
